surreal-auth-api/backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
surrealdb>=1.0.0,<2.0.0
python-jose[cryptography]>=3.3
passlib>=1.7
python-dotenv>=1.0
pydantic[email]>=2.0
pytest>=7.0
httpx>=0.27
